+++ ts/utils/supportAssistance.ts
